Added AuthNavigationController
