+++ Rocket.Chat/Controllers/Auth/AuthTableViewController.swift
@@ -146,7 +146,7 @@ class AuthTableViewController: BaseTableViewController {
     override func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)
 
-        if let nav = navigationController as? BaseNavigationController {
+        if let nav = navigationController as? AuthNavigationController {
             nav.setGrayTheme()
         }
