+++ Rocket.Chat/Controllers/Auth/LoginTableViewController.swift
@@ -145,7 +145,7 @@ class LoginTableViewController: BaseTableViewController {
     override func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)
 
-        if let nav = navigationController as? BaseNavigationController {
+        if let nav = navigationController as? AuthNavigationController {
             nav.setGrayTheme()
         }
     }
